refactor(dht)!: Use Atlas providers
Route DHT transform and inverse scaling through Hephaestus typed command streams. Borrow multidimensional Leto views, consolidate Mnemosyne scratch, preserve Moirai and Hermes execution, and exclude f64 from the f32 GPU contract at compile time.\n\nBREAKING CHANGE: DhtWgpuBackend now accepts and exposes the Hephaestus device directly, removes its raw queue accessor, and typed GPU methods require HartleyGpuStorage.\n\nRefs: backlog.md#atlas-in-house-replacement-roadmap--apollo-slice-arch

+++ crates/apollo-dht/src/infrastructure/transport/gpu/infrastructure/shaders/dht.wgsl
@@ -30,7 +30,7 @@ fn dht_transform(@builtin(global_invocation_id) gid: vec3<u32>) {
     let factor = TAU / f32(params.len);
     var sum = 0.0;
     for (var n: u32 = 0u; n < params.len; n = n + 1u) {
-        let angle = factor * f32(k * n);
+        let angle = factor * f32(k) * f32(n);
         sum = sum + input_data[n] * hartley_cas(angle);
     }
     output_data[k] = sum;
